requirements.txt
whisper>=1.0.0
openwakeword>=0.2.1
piper
transformers
torch
sounddevice
openai-whisper
mistralai>=1.6.0
python-dotenv
tflite-runtime
wave
soundfile